Refactor import scripts into own directory
Ticket: #141
